Respect disabled UltraFeed sources during fetching and ranking
